#762 - introducing the fix of changing 'float dt' to 'double dt', fixing some of the limited floating-point accuracy errors. Still need to re-run the full test suite locally, remove the comments are fulfil PEP8 linter

+++ parcels/kernel.py
@@ -250,10 +250,14 @@ class Kernel(object):
                 g.lat = g.lat.copy()
 
         fargs = [byref(f.ctypes_struct) for f in self.field_args.values()]
-        fargs += [c_float(f) for f in self.const_args.values()]
+        # fargs += [c_float(f) for f in self.const_args.values()]
+        fargs += [c_double(f) for f in self.const_args.values()]
         particle_data = pset._particle_data.ctypes.data_as(c_void_p)
         self._function(c_int(len(pset)), particle_data,
-                       c_double(endtime), c_float(dt), *fargs)
+                       c_double(endtime),
+                       c_double(dt),
+                       #c_float(dt),
+                       *fargs)
 
     def execute_python(self, pset, endtime, dt):
         """Performs the core update loop via Python"""
